支付和日历权限的功能

+++ backend/src/main/java/com/kakarote/ai_crm/service/impl/ScheduleServiceImpl.java
@@ -9,6 +9,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.kakarote.ai_crm.ai.tools.support.AiCustomerMatcher;
 import com.kakarote.ai_crm.ai.DynamicChatClientProvider;
 import com.kakarote.ai_crm.common.BasePage;
+import com.kakarote.ai_crm.common.auth.DataPermissionContext;
 import com.kakarote.ai_crm.common.exception.BusinessException;
 import com.kakarote.ai_crm.common.result.SystemCodeEnum;
 import com.kakarote.ai_crm.entity.BO.ScheduleAddBO;
@@ -20,6 +21,7 @@ import com.kakarote.ai_crm.entity.VO.ScheduleAiParseVO;
 import com.kakarote.ai_crm.entity.VO.ScheduleParticipantUserVO;
 import com.kakarote.ai_crm.entity.VO.ScheduleVO;
 import com.kakarote.ai_crm.mapper.ScheduleMapper;
+import com.kakarote.ai_crm.service.DataPermissionService;
 import com.kakarote.ai_crm.service.IGlobalSearchIndexService;
 import com.kakarote.ai_crm.service.IScheduleService;
 import com.kakarote.ai_crm.service.ManageUserService;
@@ -60,6 +62,9 @@ public class ScheduleServiceImpl extends ServiceImpl<ScheduleMapper, Schedule> i
 
     @Autowired
     private IGlobalSearchIndexService globalSearchIndexService;
+
+    @Autowired
+    private DataPermissionService dataPermissionService;
 
     @Autowired
     private AiCustomerMatcher aiCustomerMatcher;
@@ -131,6 +136,7 @@ public class ScheduleServiceImpl extends ServiceImpl<ScheduleMapper, Schedule> i
         if (ObjectUtil.isNull(schedule)) {
             throw new BusinessException(SystemCodeEnum.SYSTEM_NO_VALID, "日程不存在");
         }
+        dataPermissionService.assertUserDataAccessByPermission("schedule:delete", schedule.getCreateUserId());
         removeById(scheduleId);
         globalSearchIndexService.deleteByEntity("schedule", scheduleId);
     }
@@ -149,6 +155,9 @@ public class ScheduleServiceImpl extends ServiceImpl<ScheduleMapper, Schedule> i
     @Override
     public BasePage<ScheduleVO> queryPageList(ScheduleQueryBO queryBO) {
         queryBO.setCurrentUserId(UserUtil.getUserId());
+        DataPermissionContext context = dataPermissionService.createContextByPermission("schedule:view");
+        queryBO.setScheduleAllData(context.isAllData());
+        queryBO.setScheduleUserIds(context.getUserIds());
         BasePage<ScheduleVO> page = queryBO.parse();
         baseMapper.queryPageList(page, queryBO);
         enrichSchedules(page.getList());
